--- src/app/(service)/(responsive)/_components/mobile-top-nav.tsx
@@ -1,6 +1,0 @@
-/**
- *
- */
-export default function ResponsiveTopNav() {
-  return <></>;
-}
