+++ src/app/(service)/(모바일환경 헤더 포함)/layout.tsx
@@ -3,15 +3,19 @@ import FooterMenu from 'components/layout/footer-menu';
 import Header from 'components/layout/header';
 import ScrollToTopButton from 'components/shared/scroll-to-top-button';
 
-export default function ServiceLayout({
+export default function MyPageLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
   return (
-    <div className="flex min-h-screen flex-col">
-      {/* <Header /> */}
-      <main className="h-full w-full flex-grow">{children}</main>
+    <div className="flex min-h-screen flex-col bg-gray-50">
+      <div className="hidden lg:block">
+        <Header />
+      </div>
+      <main className="bg-background mx-auto h-full w-full max-w-[720px] flex-grow">
+        {children}
+      </main>
       <FooterMenu />
       <Footer />
