+++ src/app/(service)/(일반 헤더 - 메인페이지)/layout.tsx
