+++ src/app/globals.css
@@ -111,6 +111,6 @@ html {
 
 @media (min-width: 768px) {
   html {
-    font-size: 16px;
+    font-size: 14px;
   }
 }
