+++ src/components/layout/footer/index.tsx
@@ -7,8 +7,8 @@ const currentYear = new Date().getFullYear(); // 현재 연도 가져오기
  */
 export default function Footer() {
   return (
-    <footer className="w-full border-t border-gray-300 py-6">
-      <div className="container px-4 md:px-16">
+    <footer className="mb-15 w-full border-t border-gray-300 py-6 lg:mb-0">
+      <div className="container px-8 md:px-16">
         <Text size="xl" weight="bold">
           체험콕
         </Text>
